requirements.txt
--index-url https://pypi.tuna.tsinghua.edu.cn/simple
--trusted-host pypi.tuna.tsinghua.edu.cn

rembg[cpu]
Pillow>=9.0.0
tqdm>=4.65.0
typer>=0.9.0
rich>=13.0.0
fastapi>=0.109.0
python-multipart>=0.0.7
uvicorn>=0.27.0